requirements.txt
flask
gunicorn
opencv-python
numpy
pillow
scikit-image
face-recognition
requests
tensorflow
torch
torchvision
deepface
imagehash
